+++ lib/swaggerValidator/trafficValidator.ts
@@ -2,7 +2,6 @@ import * as fs from "fs";
 import * as path from "path";
 import { resolve as pathResolve } from "path";
 import { glob } from "glob";
-import { toLower } from "lodash";
 import {
   LiveValidationIssue,
   LiveValidator,
@@ -11,11 +10,21 @@ import {
 import { DefaultConfig } from "../util/constants";
 import { apiValidationErrors, ErrorCodeConstants } from "../util/errorDefinitions";
 import { OperationContext } from "../liveValidation/operationValidator";
+import { Options } from "../validate";
 import { traverseSwagger } from "../transform/traverseSwagger";
 import { Operation, Path, LowerHttpMethods } from "../swagger/swaggerTypes";
 import { LiveValidatorLoader } from "../liveValidation/liveValidatorLoader";
 import { inversifyGetContainer, inversifyGetInstance } from "../inversifyUtils";
+import { getApiVersionFromSwaggerPath } from "../util/utils";
 
+export interface TrafficValidationOptions extends Options {
+  sdkPackage?: string;
+  sdkLanguage?: string;
+  reportPath?: string;
+  overrideLinkInReport?: boolean;
+  specLinkPrefix?: string;
+  payloadLinkPrefix?: string;
+}
 export interface TrafficValidationIssue {
   payloadFilePath?: string;
   specFilePath?: string;
@@ -31,10 +40,17 @@ export interface RuntimeException {
 
 export interface OperationCoverageInfo {
   readonly spec: string;
+  readonly apiVersion: string;
   readonly coveredOperaions: number;
   readonly validationFailOperations: number;
+  readonly unCoveredOperations: number;
+  readonly unCoveredOperationsList: OperationMeta[];
   readonly totalOperations: number;
   readonly coverageRate: number;
+}
+
+export interface OperationMeta {
+  readonly operationId: string;
 }
 
 export class TrafficValidator {
@@ -103,7 +119,7 @@ export class TrafficValidator {
 
     const swaggerPaths = this.liveValidator.swaggerList;
     while (swaggerPaths.length > 0) {
-      let swaggerPath = swaggerPaths.shift()!;
+      const swaggerPath = swaggerPaths.shift()!;
       let spec;
       try {
         spec = await this.loader.load(pathResolve(swaggerPath));
@@ -113,7 +129,6 @@ export class TrafficValidator {
         );
       }
       if (spec !== undefined) {
-        swaggerPath = toLower(swaggerPath);
         // Get Swagger - operation mapper.
         if (this.operationSpecMapper.get(swaggerPath) === undefined) {
           this.operationSpecMapper.set(swaggerPath, []);
@@ -216,15 +231,27 @@ export class TrafficValidator {
     let coveredOperaions: number;
     let coverageRate: number;
     let validationFailOperations: number;
+    let unCoveredOperationsList: OperationMeta[];
     this.operationSpecMapper.forEach((value: string[], key: string) => {
+      unCoveredOperationsList = [];
       if (this.trafficOperation.get(key) === undefined) {
         coveredOperaions = 0;
         coverageRate = 0;
         this.coverageData.set(key, 0);
       } else if (value !== undefined && value.length !== 0) {
-        coveredOperaions = this.trafficOperation.get(key)!.length;
+        const validatedOperations = this.trafficOperation.get(key);
+        coveredOperaions = validatedOperations!.length;
         coverageRate = coveredOperaions / value.length;
         this.coverageData.set(key, coverageRate);
+        const unValidatedOperations = [...value];
+        validatedOperations!.forEach((element) => {
+          unValidatedOperations.splice(unValidatedOperations.indexOf(element), 1);
+        });
+        unValidatedOperations.forEach((element) => {
+          unCoveredOperationsList.push({
+            operationId: element,
+          });
+        });
       } else {
         coveredOperaions = 0;
         coverageRate = 0;
@@ -236,12 +263,26 @@ export class TrafficValidator {
       } else {
         validationFailOperations = this.validationFailOperations.get(key)!.length;
       }
+      const sortedUnCoveredOperationsList = unCoveredOperationsList.sort(function (op1, op2) {
+        const opId1 = op1.operationId;
+        const opId2 = op2.operationId;
+        if (opId1 < opId2) {
+          return -1;
+        }
+        if (opId1 > opId2) {
+          return 1;
+        }
+        return 0;
+      });
       this.operationCoverageResult.push({
         spec: key,
+        apiVersion: getApiVersionFromSwaggerPath(key),
         coveredOperaions: coveredOperaions,
         coverageRate: coverageRate,
+        unCoveredOperations: value.length - coveredOperaions,
         totalOperations: value.length,
         validationFailOperations: validationFailOperations,
+        unCoveredOperationsList: sortedUnCoveredOperationsList,
       });
     });
     return this.trafficValidationResult;
@@ -255,8 +296,9 @@ export class TrafficValidator {
     for (const key of this.operationSpecMapper.keys()) {
       const value = this.operationSpecMapper.get(key);
       if (
-        key.includes(toLower(operationInfo.apiVersion)) &&
-        key.includes(toLower(operationInfo.validationRequest?.providerNamespace))
+        key.includes(operationInfo.validationRequest?.providerNamespace) &&
+        (key.includes(operationInfo.apiVersion) ||
+          key.toLowerCase().includes(operationInfo.apiVersion))
       ) {
         if (value!.includes(operationInfo.operationId)) {
           result = key;
@@ -272,8 +314,9 @@ export class TrafficValidator {
     for (const key of this.operationSpecMapper.keys()) {
       const value = this.operationSpecMapper.get(key);
       if (
-        key.includes(toLower(operationInfo.apiVersion)) &&
-        value!.includes(operationInfo.operationId)
+        value!.includes(operationInfo.operationId) &&
+        (key.includes(operationInfo.apiVersion) ||
+          key.toLowerCase().includes(operationInfo.apiVersion))
       ) {
         result = key;
         return result;